Move Setup of VeConcatRDD into a factory method
