+++ src/main/scala/com/nec/ve/VeConcatRDD.scala
@@ -1,6 +1,7 @@
 package com.nec.ve
 
-import com.nec.native.CompiledVeFunction
+import com.nec.native.{CompiledVeFunction, FunctionTyping}
+import com.nec.spark.agile.merge.MergeFunction
 import com.nec.ve.colvector.VeColBatch.VeBatchOfBatches
 import org.apache.spark.rdd.RDD
 
@@ -23,5 +24,23 @@ class VeConcatRDD[U: TypeTag, T: TypeTag](
         Iterator(VeColBatch.fromList(res))
       }
     }
+  }
+}
+
+object VeConcatRDD {
+  def apply[U: TypeTag, T: TypeTag](rdd: RDD[VeColBatch], funcTypes: FunctionTyping[T, T]): VeConcatRDD[U, VeColBatch] = {
+    import com.nec.native.SyntaxTreeOps._
+
+    val dataType = funcTypes.output.tpe.toVeType
+
+    val funcName = s"merge_${dataType.toString}_1"
+    val code = MergeFunction(funcName, List(dataType))
+    val func = CompiledVeFunction(
+      code.toCFunction,
+      code.toVeFunction.namedResults,
+      funcTypes
+    )
+
+    new VeConcatRDD[U, VeColBatch](new RawVeRDD[T](rdd), func)
   }
 }
